+++ src/test/java/pokeraidbot/MiscTests.java
@@ -11,5 +11,10 @@ public class MiscTests {
     public void testRegExpForPlusCommand() throws Exception {
         Assert.assertThat("+1 10:35 Hästen".matches(plusXRegExp),
                 is(true));
+        Assert.assertThat("+1 anmälning känns vettigt att ha i chatten".matches(plusXRegExp),
+                is(false));
+
+        Assert.assertThat("+1 Hästen 10:35".matches(plusXRegExp),
+                is(false));
     }
 }